Major Update
[New Features:]
Added secondary page for releases that are under album type (click on album name to redirect to).
Added config.conf file.

[Changes:]
Redid folder structure for database.
Redid image naming scheme for all releases.
Redid PHP scripts to make it slightly easier to read.
Removed Discogs ID: ###### link for albums as this is now present on the secondary page that opens when you click on the albums name.

[Bug Fixes:]
Fixed the stats php scripts which did not take into consideration that some of the database files do not have actual entries starting on line 0.